Correct UI of products and inventory tabs
Add check availability button in the add and edit product modals, beside barcode field, in the products tab so as to check whether the barcode is available or not.
Add search button in add and edit inventory modals, beside barcode field, in the inventory tab so as to see whether the barcode is there or not.
Ensure consistency in all the buttons of brands, products, inventory and orders tabs.